chore: use @node-llm as model backend
